refactor: move "Files" resource to separate project (#34)
