+++ VirusTotalCore.sln
@@ -12,6 +12,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VirusTotalCore.IpAddresses"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VirusTotalCore.Urls", "src\VirusTotalCore.Urls\VirusTotalCore.Urls.csproj", "{A1F2E3D4-5B6C-7D8E-9F0A-1B2C3D4E5F6A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VirusTotalCore.Files", "src\VirusTotalCore.Files\VirusTotalCore.Files.csproj", "{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -82,6 +84,18 @@ Global
 		{A1F2E3D4-5B6C-7D8E-9F0A-1B2C3D4E5F6A}.Release|x64.Build.0 = Release|Any CPU
 		{A1F2E3D4-5B6C-7D8E-9F0A-1B2C3D4E5F6A}.Release|x86.ActiveCfg = Release|Any CPU
 		{A1F2E3D4-5B6C-7D8E-9F0A-1B2C3D4E5F6A}.Release|x86.Build.0 = Release|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Debug|x64.Build.0 = Debug|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Debug|x86.Build.0 = Debug|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Release|x64.ActiveCfg = Release|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Release|x64.Build.0 = Release|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Release|x86.ActiveCfg = Release|Any CPU
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -90,5 +104,6 @@ Global
 		{1BB3E737-DDE4-4424-A0E5-BB8AA050E37A} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{9E32E4C2-4BC4-4277-B7F4-7A250CEC6B99} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{A1F2E3D4-5B6C-7D8E-9F0A-1B2C3D4E5F6A} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{B3C4D5E6-7F8A-9B0C-1D2E-3F4A5B6C7D8E} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 EndGlobal
